pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"


[project]
name = "opt_guarantees"
description="A package for warm-starting optimization algorithms, using data."
readme = "README.md"
license = {text = "Apache 2.0"}
dynamic = ["version"]
authors = [
{name = "Rajiv Sambharya", email="rajivs@princeton.edu"},
{name = "Bartolomeo Stellato", email="bstellato@princeton.edu"}
]
dependencies = [
    "numpy",
    "scipy",
    "jax==0.4.6",
    "jaxlib==0.4.14",
    "cvxpy>=1.3.0",
    "matplotlib",
    "jaxopt",
    "optax==0.1.5",
    "matplotlib",
    "hydra-core", 
    "emnist"
]

[tool.setuptools.packages.find]
include = ["opt_guarantees*"]
exclude = ["tutorials*", "benchmarks*", "tests*"]

[tool.setuptools_scm]
# To infer version automatically from git
write_to = "opt_guarantees/_version.py"

[project.optional-dependencies]
dev = ["pytest", "ruff", "ruff-lsp", "black", "pandas", "jupyterlab"]


[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
select = ["E", "F", "I"]
ignore = ["E722"]
line-length = 100
exclude = ["build", "examples", "instances", "docs", "*__init__.py"]
target-version = "py310"
